requirements.txt
asn1crypto==0.24.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
cached-property==1.3.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.2.2
docker==2.5.1
docker-compose==1.17.1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
ez-setup==0.9
Flask==1.0.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
line-bot-sdk==1.8.0
lxml==4.2.1
MarkupSafe==1.0
mechanize==0.3.6
mock==2.0.0
mysql-connector-python-rf==2.2.2
mysqlclient==1.3.10
oauth2client==4.1.3
pbr==3.1.1
pipreqs==0.4.9
powerline-status==2.6
psutil==5.4.2
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.6.4
pyOpenSSL==18.0.0
pyperclip==1.6.0
pytz==2018.4
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
rsa==4.0
SecretStorage==2.3.1
selenium==3.9.0
Send2Trash==1.5.0
six==1.11.0
SQLAlchemy==1.2.12
texttable==0.9.1
twilio==6.14.4
unity-lens-photos==1.0
urllib3==1.22
webencodings==0.5.1
websocket-client==0.44.0
Werkzeug==0.14.1
yarg==0.1.9
